refactor: inline Identity::create into IdentityManager::create_new_identity
